[#1770] fstree: Depth parameter `int` -> `uint64`
Negative values have no sense. On the other hand it differs from the
blobovnicza's configuration and prevents unification.

+++ pkg/local_object_storage/blobstor/fstree/option.go
@@ -6,7 +6,7 @@ import (
 
 type Option func(*FSTree)
 
-func WithDepth(d int) Option {
+func WithDepth(d uint64) Option {
 	return func(f *FSTree) {
 		f.Depth = d
 	}
